Update dependencies and convert to double

+++ Drivers/DriverBase/Interfaces/IPressureSensor.cs
@@ -6,19 +6,19 @@
         /// Read pressure from a sensor, raw value
         /// </summary>
         /// <returns>Pressure in RAW value</returns>
-        float ReadPressure();
+        double ReadPressure();
         /// <summary>
         /// Calculates pressure from raw to target type
         /// </summary>
         /// <param name="type">Pressure type</param>
         /// <param name="rawPressure">Raw pressure value</param>
         /// <returns>Target unit pressure from sensor</returns>
-        float CalculatePressure(Enums.PressureType type, float rawPressure);
+        double CalculatePressure(Enums.PressureType type, double rawPressure);
         /// <summary>
         /// Reads pressure from sensor and calculates pressure to your unit
         /// </summary>
         /// <param name="type">Pressure unit type</param>
         /// <returns>Target unit pressure from sensor</returns>
-        float ReadPressure(Enums.PressureType type);
+        double ReadPressure(Enums.PressureType type);
     }
 }
